// src/app/model/patient.ts
export class Patient {
    id:number;
    fullName:String;
    phone:String;
    isActive: number;
    dob: Date;
    reason: String
    weight: number;
    gender: number;
    job: String;
    jobPlace: String;
    email: String;
    bloodType: String
    height: number;
    streetName: String;
    imageUrl: String;
    ethnicName: String;
    districtName: String;
    wardName: String;
    cityName: String;
}